fix(quiz-sync): address subagent review findings on PR #1487
A code-reviewer subagent surfaced four issues against the deferred-
followups PR. All four are addressed here.

1. CRITICAL: Rejoin path didn't reset preSyncVersion.
   When a student under the attempt cap rejoins after completing,
   `useQuizSession`'s reset (status/answers/score/submittedAt)
   missed `preSyncVersion`. A response previously tagged at e.g.
   `preSyncVersion: 4` would carry that chip onto a fresh attempt
   taken against post-sync content, AND be invisible to future
   `where('preSyncVersion', '==', 0)` queries. The rejoin reset now
   clears it to 0, and the firestore rule's student-update branch
   admits `preSyncVersion` to the `hasOnly` list with a constraint
   that the value must be 0 (mirrors the create-time guard so a
   forged client can't dodge sync tagging).

2. IMPORTANT: previousSyncedVersion-of-0 sentinel collision in
   syncAssignmentToLatest. If `assignment.sync.syncedVersion` ever
   landed at 0 (canonical-init change, corrupt doc), the tag write
   would be `preSyncVersion: 0` — the same value the server query
   targets. Subsequent syncs would re-tag the same rows forever,
   causing write amplification. Fixed by flooring the tag value at 1
   (`Math.max(previousSyncedVersion, 1)`) so the tag never collides
   with the untagged sentinel.

3. IMPORTANT: shareAssignment skipped the sync-linkage migrator.
   Only `migrateLegacyAssignmentShape` was applied — the only
   assignment-read site post-refactor that broke the "every read goes
   through both migrators" invariant. Function doesn't currently read
   `assignment.sync`, so no live bug, but a future change reading
   `.sync.groupId` here would silently fail on legacy docs. Now
   pipes through both migrators to match every other read site.

4. SUGGESTION: Duplicated migrator logic across hooks consolidated.
   `migrateQuizMetadataShape` (useQuiz.ts) and
   `migrateQuizMetadataShapeFromUnknown` (useQuizAssignments.ts) had
   already drifted on validation strictness. Lifted to
   `utils/quizSyncMigration.ts` so the same rules drive every read
   site — exactly the failure mode the abstraction prevents.
